Removed extended warnings
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@14846 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: 0785bca09f0a1d09b88d810001668bea3797fcce

+++ core/rcp/de.uniol.inf.is.odysseus.rcp.dashboard.test/src/de/uniol/inf/is/odysseus/rcp/dashboard/test/ConfigurationTest.java
@@ -58,15 +58,15 @@ public class ConfigurationTest {
 
 	@Test
 	public void testGetAndSet() {
-		assertEquals(configuration.get("IntSetting"), (Integer)100);
+		assertEquals(configuration.get("IntSetting"), 100);
 		
 		configuration.set("IntSetting", 200);
 		
-		assertEquals(configuration.get("IntSetting"), (Integer)200);
+		assertEquals(configuration.get("IntSetting"), 200);
 		
 		configuration.reset("IntSetting");
 		
-		assertEquals(configuration.get("IntSetting"), (Integer)100);
+		assertEquals(configuration.get("IntSetting"), 100);
 	}
 	
 	@Test
